split query plan rearangement into separate file
